Implementing Workout Service

+++ src/main/java/ednax/dio/santander/restapi/models/ExerciseModel.java
@@ -1,7 +1,5 @@
 package ednax.dio.santander.restapi.models;
 
-import java.util.UUID;
-
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -21,7 +19,7 @@ public class ExerciseModel {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(length = 50, nullable = false)
     private String name;
